move handlers to src directly remove not needed app.eventFilter document registering handlers
