+++ src/CustomLineDrawHandler.cpp
@@ -17,45 +17,70 @@
  *   59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.             *
  ***************************************************************************/
 
-#include "map/handlers/RoomMoveActivationHandler.h"
+#include "CustomLineDrawHandler.h"
+
+#include "TMap.h"
+#include "TRoom.h"
+#include "TRoomDB.h"
 
 #include "pre_guard.h"
+#include <QEvent>
 #include <QMouseEvent>
-#include <QRect>
+#include <QPointF>
 #include "post_guard.h"
 
-RoomMoveActivationHandler::RoomMoveActivationHandler(T2DMap& mapWidget)
+CustomLineDrawHandler::CustomLineDrawHandler(T2DMap& mapWidget)
 : mMapWidget(mapWidget)
 {
 }
 
-bool RoomMoveActivationHandler::matches(const T2DMap::MapInteractionContext& context) const
+bool CustomLineDrawHandler::matches(const T2DMap::MapInteractionContext& context) const
 {
-    if (!context.event) {
+    if (!context.event || !mMapWidget.mpMap) {
         return false;
     }
 
-    if (!context.isRoomBeingMoved) {
+    if (context.event->type() != QEvent::MouseButtonPress) {
         return false;
     }
 
-    return context.event->type() == QEvent::MouseButtonPress
-        && context.button == Qt::LeftButton;
+    if (context.button != Qt::LeftButton) {
+        return false;
+    }
+
+    if (!context.isCustomLineDrawing) {
+        return false;
+    }
+
+    if (context.customLinesRoomFrom <= 0) {
+        return false;
+    }
+
+    return true;
 }
 
-bool RoomMoveActivationHandler::handle(T2DMap::MapInteractionContext& context)
+bool CustomLineDrawHandler::handle(T2DMap::MapInteractionContext& context)
 {
-    if (!context.event || !context.isRoomBeingMoved) {
+    if (!mMapWidget.mpMap || !mMapWidget.mpMap->mpRoomDB) {
         return false;
     }
 
-    mMapWidget.mPopupMenu = false;
-    mMapWidget.mPick = true;
-    mMapWidget.setMouseTracking(false);
-    mMapWidget.mRoomBeingMoved = false;
-    mMapWidget.mMultiRect = QRect(0, 0, 0, 0);
+    if (context.isDialogLocked) {
+        return true;
+    }
 
-    context.isRoomBeingMoved = false;
+    TRoom* room = mMapWidget.mpMap->mpRoomDB->getRoom(context.customLinesRoomFrom);
 
-    return false;
+    if (!room) {
+        return false;
+    }
+
+    const float mapX = static_cast<float>(context.mapX);
+    const float mapY = static_cast<float>(context.mapY);
+
+    room->customLines[context.customLinesRoomExit].push_back(QPointF(mapX, mapY));
+    room->calcRoomDimensions();
+    mMapWidget.repaint();
+
+    return true;
 }
