+++ src/CustomLineEditContextMenuHandler.h
