+++ src/LabelInteractionHandler.h
