+++ src/PanInteractionHandler.h
