+++ src/RoomContextMenuHandler.cpp
@@ -17,7 +17,7 @@
  *   59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.             *
  ***************************************************************************/
 
-#include "map/handlers/RoomContextMenuHandler.h"
+#include "RoomContextMenuHandler.h"
 
 #include "TArea.h"
 #include "TMap.h"
