+++ src/RoomMoveActivationHandler.cpp
@@ -17,70 +17,45 @@
  *   59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.             *
  ***************************************************************************/
 
-#include "map/handlers/CustomLineDrawHandler.h"
-
-#include "TMap.h"
-#include "TRoom.h"
-#include "TRoomDB.h"
+#include "RoomMoveActivationHandler.h"
 
 #include "pre_guard.h"
-#include <QEvent>
 #include <QMouseEvent>
-#include <QPointF>
+#include <QRect>
 #include "post_guard.h"
 
-CustomLineDrawHandler::CustomLineDrawHandler(T2DMap& mapWidget)
+RoomMoveActivationHandler::RoomMoveActivationHandler(T2DMap& mapWidget)
 : mMapWidget(mapWidget)
 {
 }
 
-bool CustomLineDrawHandler::matches(const T2DMap::MapInteractionContext& context) const
+bool RoomMoveActivationHandler::matches(const T2DMap::MapInteractionContext& context) const
 {
-    if (!context.event || !mMapWidget.mpMap) {
+    if (!context.event) {
         return false;
     }
 
-    if (context.event->type() != QEvent::MouseButtonPress) {
+    if (!context.isRoomBeingMoved) {
         return false;
     }
 
-    if (context.button != Qt::LeftButton) {
-        return false;
-    }
-
-    if (!context.isCustomLineDrawing) {
-        return false;
-    }
-
-    if (context.customLinesRoomFrom <= 0) {
-        return false;
-    }
-
-    return true;
+    return context.event->type() == QEvent::MouseButtonPress
+        && context.button == Qt::LeftButton;
 }
 
-bool CustomLineDrawHandler::handle(T2DMap::MapInteractionContext& context)
+bool RoomMoveActivationHandler::handle(T2DMap::MapInteractionContext& context)
 {
-    if (!mMapWidget.mpMap || !mMapWidget.mpMap->mpRoomDB) {
+    if (!context.event || !context.isRoomBeingMoved) {
         return false;
     }
 
-    if (context.isDialogLocked) {
-        return true;
-    }
+    mMapWidget.mPopupMenu = false;
+    mMapWidget.mPick = true;
+    mMapWidget.setMouseTracking(false);
+    mMapWidget.mRoomBeingMoved = false;
+    mMapWidget.mMultiRect = QRect(0, 0, 0, 0);
 
-    TRoom* room = mMapWidget.mpMap->mpRoomDB->getRoom(context.customLinesRoomFrom);
+    context.isRoomBeingMoved = false;
 
-    if (!room) {
-        return false;
-    }
-
-    const float mapX = static_cast<float>(context.mapX);
-    const float mapY = static_cast<float>(context.mapY);
-
-    room->customLines[context.customLinesRoomExit].push_back(QPointF(mapX, mapY));
-    room->calcRoomDimensions();
-    mMapWidget.repaint();
-
-    return true;
+    return false;
 }
