+++ src/RoomMoveActivationHandler.h
