+++ src/RoomMoveDragHandler.h
