+++ src/SelectionRectangleHandler.cpp
@@ -17,7 +17,7 @@
  *   59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.             *
  ***************************************************************************/
 
-#include "map/handlers/SelectionRectangleHandler.h"
+#include "SelectionRectangleHandler.h"
 
 #include "TArea.h"
 #include "TMap.h"
