+++ src/SelectionRectangleHandler.h
